refactoring/remove models in js directory
