+++ js/change-state.js
@@ -1,7 +1,7 @@
 import renderArtist from "./screens/artist-view-screen";
 import renderGenre from "./screens/genre-view-screen";
 import renderSuccess from "./screens/succes-view-screen";
-import {game} from "./data/models/game";
+import {game} from "./models/game";
 
 const changeView = (state) => {
   if (state.level < game.levels.length) {
