+++ js/models/initialization-state.js
@@ -1,4 +1,4 @@
-import config from "./../../lib/config";
+import config from "../lib/config";
 
 const initializationState = () => {
   return Object.assign({}, config);
