+++ js/screens/welcome-view-screen.js
@@ -1,8 +1,8 @@
-import WelcomeView from "../view/welcome-view";
-import {startGame} from "./../data/models/startGame";
+import WelcomeView from "./../view/welcome-view";
+import {startGame} from "./../models/startGame";
 import {renderTemplate} from "./../utils";
-import {initializationState} from "./../data/models/initialization-state";
-import changeView from "../change-state";
+import {initializationState} from "./../models/initialization-state";
+import changeView from "./../change-state";
 
 const renderWelcome = () => {
   const previewView = new WelcomeView(startGame);
